Bug 1918287: pass ovirtClient to identity and remove redundant call to Test connection
We didn't pass the ovirtClient to the identity on creation which caused a new connection to be created on each call to ovirtClient.GetConnection().
On top of that we called the Test function which is redundant since it is either a new connection or it is being called in the function.

Those 2 errors caused the logs to get filled with auth requests and the evnet tab in the engine to get filled with auth events.