node: implement NodeGetVolumeStats capability
Adds a missing capability to Node plugin which reports volume stats
to users.

There are two cases which can occur:

1) volume is block device
We can not obtain full usage but report at least total capacity of the
device.

2) volume has a filesystem
After we rule out block device we assume filesystem presence and can
provide full stats.

Alternative way would be using ovirt engine API to get the stats but
we can not be certain the size is reported correctly because the volume
can be a RAW image which when compared to QCOW lacks metadata resulting
in possibly incorrect usage report.